feat: input box 너비 설정
