+++ app/my-profile/edit/page.module.scss
@@ -4,10 +4,10 @@
   padding-top: 6rem;
   max-width: 96.4rem;
   margin: 0 auto;
-  margin-bottom: 3.2rem;
-
+  
   
   header {
+    margin-bottom: 3.2rem;
     @include flexbox(row, space-between, center);
     span {
       font-size: 2.8rem;
