+++ app/my-profile/edit/page.tsx
@@ -1,5 +1,6 @@
 import { EditProfile } from "components/employee";
 import Image from "next/image";
+import Link from "next/link";
 import React from "react";
 import styles from "./page.module.scss";
 
@@ -8,7 +9,9 @@ const page = () => {
     <div className={styles.layout}>
       <header>
         <span>내 프로필</span>
-        <Image src="/images/close.svg" alt="닫기 버튼" width={30} height={30} />
+        <Link href="/my-profile">
+          <Image src="/images/close.svg" alt="닫기 버튼" width={30} height={30} />
+        </Link>
       </header>
       <EditProfile />
     </div>
