+++ components/common/CustomInput/CustomInput.module.scss
@@ -1,4 +1,5 @@
 .box {
+  width: 100%;
   height: 11.6rem;
 
   .label {
