+++ components/employee/EditProfile/EditProfile.module.scss
@@ -1,0 +1,22 @@
+.layout {
+  .inputBox {
+    display: flex;
+    gap: 2rem;
+    
+    .box {
+      flex-grow: 1;
+    }
+  }
+
+  textarea {
+    width: 100%;
+    height: 15.3rem;
+    resize: none;
+  }
+
+  button {
+    display: block;
+    width: 31.2rem;
+    margin: 10rem auto 0;
+  }
+}
